* added randomcolor package to generate anonymous author colors * fixed bug where region borders did not match author.color * useLayoutEffect instead of useEffect to call setMessage (to avoid flashing the wrong author.color when mounting SemanticScreen) * move initialState.ts to constants since it only gets imported in App.tsx, not in any reducers
