+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^7.1.2",
     "@types/jest": "^24.0.0",
     "@types/node": "^12.0.0",
+    "@types/randomcolor": "^0.5.5",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-redux": "^7.1.9",
@@ -16,6 +17,7 @@
     "animate-css-grid": "^1.4.3",
     "immutability-helper": "^3.1.1",
     "prettier": "^2.0.5",
+    "randomcolor": "^0.6.2",
     "react": "^16.13.1",
     "react-dnd": "^11.1.3",
     "react-dnd-html5-backend": "^11.1.3",
