+++ src/constants/initialState.ts
@@ -18,15 +18,7 @@
 */
 //TODO: how to confirm that focus point exists in the array of points
 //contained in the semscreen?
-import { AuthorI } from "../dataModels";
 import { MessageState } from "../reducers/message";
-
-export const authors: AuthorI[] = [
-  {
-    name: "KindWoman",
-    color: "#7d3989",
-  },
-];
 
 export const messages: MessageState[] = [
   {
